Unify status bar menus for copilot and supermaven
